Backend/requirements.txt
azure-ai-inference
azure-core
python-dotenv
flask
gunicorn
flask-cors

# New dependencies for FastAPI
fastapi>=0.68.0,<1.0.0
uvicorn>=0.15.0,<1.0.0
requests>=2.26.0,<3.0.0
python-multipart>=0.0.5,<1.0.0